multi-container-app_Assignment4/python-app/requirements.txt
# requirements.txt
Flask>=2.2
psycopg2-binary==2.9.1
redis==4.1.1
